[IMP] attribute journals to their chart
At current it seems that journal creation through chart definition has been abandoned, although the creator of the `generate_journal` method has contemplated this possibility through `journals_dict=None` signature. As we grow towards industry specific accounting verticals, it would be very useful to incorporate this functionality again in the future. Meanwhile, we did so in private code and stumbled over a inconsistency: The journal creation is not inline with the other `generate_` methods in that it doesn't use the xml creation, nor the external id `_update` method. Thanks to this invention, creating an individual update wizard is easy, yet journals fall out of the line and need an override.
This commits puts journals in line, so they stand all nicely in a row.